src/emojiforge_backend/requirements.txt
torch>=2.4.0
torchvision
torchaudio
--extra-index-url https://download.pytorch.org/whl/cu128
diffusers @ git+https://github.com/huggingface/diffusers.git
transformers>=4.51.0,<5
accelerate>=1.0.0
safetensors
huggingface_hub
Pillow>=10.0.0
emoji>=2.12.0
bitsandbytes>=0.45.0
sentencepiece
protobuf
rembg>=2.0.50
